Don't use compiled headers to check padding
Using compiled headers as a means of checking our exported headers for
padding was a silly idea.  It seemed convenient, but the existance of
the compiled header files triggers their automatic use by gcc which is
very much not appreciated.  Instead just do the simple thing of
compiling them into an unused object file with -Wpadded.

Signed-off-by: Zach Brown <zab@versity.com>