Update block btree
There's a lot of evolution of the block btree in here.

We move to a fixed size key that has 3 u64s, mostly to simplify
management of the keys.  We add a small number of key functions for
setting, comparing, incrementing, etc.

We sort the items in the block with a persistent rbtree instead of the
array of item offsets.  This makes insertion a lot cheaper because we
only need to initialize a new item and link it into the tree at the
leaf.  This change impacts nearly all of the code as everything needs to
work with items and nodes instead of array positions.

The interface only provides iterators for working with items.  This
obviously benefits iterating callers like readdir and writers that want
to merge sets of items.  We may want to provide some simple one-off
wrappers in terms of the iterators.

Signed-off-by: Zach Brown <zab@versity.com>