Update block cache for txn getting refs
The previous block transaction and caching model had the transaction
core prepare and perform a set of atomic modifications.  Modifications
had to be able to build up and understand the full setof blocks and the
changes they'd want to make when building a transation.  This is a bit
too much to ask of complicated operations that modify multiple items in
a one btree (renaming dirents).

This shifts the block cache to serve the changes to the transaction core
which have it incrementally build up references to blocks, modifying
contents as it goes.

We add the notion of shared read and exclusive write references.  We add
an atomic which manages exclusion and block on it when acquiring
references.  We exclude write references while dirty blocks are being
written by having them hold a read reference, which still allows shared
reads.

Fundeamental to the new model for building up dirty blocks in a
transaction is the support for backing off and undoing the transaction
before retrying.  We only mark blocks dirty as write references are put,
which gives the transaction the chance to stop us from dirtying the
block as it retries.

We rename the writeback work to flush work, which at least makes more
sense to me.  We add a clean worker and list which lets us implement
sync by injecting fake blocks into the ordered flush->clean lists.  The
sync waiter is woken by the cleaner once all the dirty blocks they were
waiting for have been written.

We get rid of the sets that tracked dependent changes to blocks.  This
was reasonably easy to do when the transaction had prepared the set of
blocks but is a lot more complicated when blocks are dirtied
incrementally.  We'll save that for another day.

Signed-off-by: Zach Brown <zab@versity.com>