Flesh out userspace tracing
The previous tracing core was a bit too heavy.  It was influenced by
having built other subsystems that passed data around between threads.
Specfically, it was built around the idea of passing trace buffers to
and from, say, threads that were sending them over the network to
readers that were gathering traces from a lot of processes.

It was all a bit much.  This change makes it a simpler at it's core.
Each thread gets a static ring buffer.  We use the pthread
thread-local-storage calls directly intead of the sketchy urcu per-cpu
wrappers.  This cleans up setup and teardown of the tracing a bit, and
we take the opportunity to return errors from thread creation if
per-thread setup fails.

The trace buffers are also built in files so that they persist after
process crashes.  This gets rid of the flushing code entirely.

We also add cheap timestamps.  Modern Intel architectures have TSCs that
advance at a constant rate and have unpriviledged instructions to learn
that rate.  We add trace file headers that record that rate and store a
synchronization point between the tracing ticks and the realtime wall
clock.

There's still of work to do to make this more robust, but this is a
solid start to let us use tracing in initial debugging.

Signed-off-by: Zach Brown <zab@versity.com>